+++ src/List.hs
@@ -6,6 +6,8 @@ module List
        , ordNub
        , sortOn
        , list
+       , product
+       , sum
        , unzip
        , unzip3
        , zip
@@ -13,13 +15,14 @@ module List
        ) where
 
 import           Control.Monad (return)
-import           Data.Foldable (Foldable, foldr)
+import           Data.Foldable (Foldable, foldl', foldr)
 import           Data.Function ((.))
 import           Data.Functor  (fmap)
 import           Data.List     (sortBy, unzip, unzip3, zip, zip3)
 import           Data.Maybe    (Maybe (..))
 import           Data.Ord      (Ord, comparing)
 import qualified Data.Set      as Set
+import           GHC.Num       (Num, (*), (+))
 
 head :: (Foldable f) => f a -> Maybe a
 head = foldr (\x _ -> return x) Nothing
@@ -41,3 +44,11 @@ list :: [b] -> (a -> b) -> [a] -> [b]
 list def f xs = case xs of
   [] -> def
   _  -> fmap f xs
+
+{-# INLINE product #-}
+product :: (Foldable f, Num a) => f a -> a
+product = foldl' (*) 1
+
+{-# INLINE sum #-}
+sum :: (Foldable f, Num a) => f a -> a
+sum = foldl' (+) 0